Export VERSION information (#794)
* Export VERSION information

As pointed out by @networkimprov, it is a good practice to tell the
package version by exporting a constant or having a comment to
indicate it. Therefere, the version info is exported now.

Resolves #747.

* Revert "Export VERSION information"

This reverts commit 972e15758f0cdcd33d6ebe7cfd1fc7537f6355d1.

@GillesDebunne and I agreed on a different approach, which is based
on a hardcoded VERSION export, while maintaining a logic to ensure
that the VERSION variable matches the package version.

See https://github.com/moment/luxon/pull/800#issuecomment-703887163.

* Export hardcoded VERSION variable

For a less bold approach, the exported VERSION variable is now
hardcoded. Meanwhile, the `scripts/tag` script was improved to
ensure that all versions match.

See https://github.com/moment/luxon/pull/800#issuecomment-703887163.

Resolves #747 and resolves #799.

* Move version validation to dedicated script

Move the version validation to a dedicated script, so that it can
be called both from release and tag scripts, which are scripts that
should validate the version before proceeding.